# Create indices for bookshelves_events table