Noticing ~1/3 of time of `/search` spent in db queries according to sentry profiles ([link](https://sentry.archive.org/organizations/ia-ux/performance/summary/profiles/?environment=production&project=7&statsPeriod=24h&transaction=%2Fsearch)):
![image](https://github.com/user-attachments/assets/2e3f5e61-4da8-47a4-b60d-cbd691e5ce8d)

I created 50k dummy data rows locally, and ran the 3 types of queries we run against this table. The one multi-index seems to have removed all Seq Scans and drastically dropped down the query time!

| Query | Time Before | Time After |
| ---- | ---- | ---- |
| Query 1 | 124.643 ms | 5.099 ms |
| Query 2 | 44.402 ms | 1.072 ms |
| Query 3 | 47.150 ms | 0.108 ms |

It might be worth experimenting with adding more indices, but this seems sufficient at the moment. And it would be good to run the below queries on ol-db1 as well before/after.


```sql
-- Sample check-in button query
EXPLAIN ANALYZE SELECT id, event_date FROM bookshelves_events
WHERE username='user_42' AND work_id IN (1234, 143)
AND event_type=3
ORDER BY event_date DESC LIMIT 1;

-- Sample filtered reading log query
EXPLAIN ANALYZE SELECT b.work_id, b.created, b.edition_id
FROM bookshelves_books b
INNER JOIN bookshelves_events e
ON b.work_id = e.work_id AND b.username = e.username
WHERE b.username = 'user_42'
AND e.event_date LIKE '2024%'
ORDER BY b.created DESC;

-- Find all books with events for a given user
EXPLAIN ANALYZE SELECT DISTINCT ON (work_id) work_id, * FROM bookshelves_events
WHERE username='user_42'
    AND event_type=3
    AND event_date LIKE '2024%'
ORDER BY work_id, updated DESC;
```

## Results without index:

Query 1:
```
                                                          QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------   
 Limit  (cost=1961.71..1961.71 rows=1 width=9) (actual time=120.035..120.035 rows=0 loops=1)
   ->  Sort  (cost=1961.71..1961.71 rows=1 width=9) (actual time=118.147..118.147 rows=0 loops=1)
         Sort Key: event_date
         Sort Method: quicksort  Memory: 25kB
         ->  Seq Scan on bookshelves_events  (cost=0.00..1961.70 rows=1 width=9) (actual time=114.946..114.946 rows=0 loops=1)    
               Filter: ((work_id = ANY ('{1234,143}'::integer[])) AND (username = 'user_42'::text) AND (event_type = 3))
               Rows Removed by Filter: 70497
 Total runtime: 124.643 ms
```

Query 2:
```
                                                                       QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------------------------
 Sort  (cost=1793.91..1793.92 rows=1 width=16) (actual time=44.339..44.339 rows=0 loops=1)
   Sort Key: b.created
   Sort Method: quicksort  Memory: 25kB
   ->  Nested Loop  (cost=0.41..1793.90 rows=1 width=16) (actual time=44.308..44.308 rows=0 loops=1)
         ->  Seq Scan on bookshelves_events e  (cost=0.00..1785.45 rows=1 width=13) (actual time=39.111..44.266 rows=1 loops=1)
               Filter: ((event_date ~~ '2024%'::text) AND (username = 'user_42'::text))
               Rows Removed by Filter: 70496
         ->  Index Scan using bookshelves_books_pkey on bookshelves_books b  (cost=0.41..8.44 rows=1 width=25) (actual time=0.032..0.032 rows=0 loops=1)
               Index Cond: ((username = 'user_42'::text) AND (work_id = e.work_id))
 Total runtime: 44.402 ms
```

Query 3:
```
                                                          QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------
 Unique  (cost=1961.71..1961.72 rows=1 width=78) (actual time=47.110..47.111 rows=1 loops=1)
   ->  Sort  (cost=1961.71..1961.71 rows=1 width=78) (actual time=47.108..47.108 rows=1 loops=1)
         Sort Key: work_id, updated
         Sort Method: quicksort  Memory: 25kB
         ->  Seq Scan on bookshelves_events  (cost=0.00..1961.70 rows=1 width=78) (actual time=36.892..45.651 rows=1 loops=1)
               Filter: ((event_date ~~ '2024%'::text) AND (username = 'user_42'::text) AND (event_type = 3))
               Rows Removed by Filter: 70496
 Total runtime: 47.150 ms
```

## Results after creating multi-key index:

Query 1:

```
                                                                              QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=12.88..12.89 rows=1 width=9) (actual time=0.301..0.301 rows=0 loops=1)
   ->  Sort  (cost=12.88..12.89 rows=1 width=9) (actual time=0.299..0.299 rows=0 loops=1)
         Sort Key: event_date
         Sort Method: quicksort  Memory: 25kB
         ->  Index Scan using bookshelves_events_user_checkins_idx on bookshelves_events  (cost=0.42..12.87 rows=1 width=9) (actual time=0.288..0.288 rows=0 loops=1)
               Index Cond: ((username = 'user_42'::text) AND (work_id = ANY ('{1234,143}'::integer[])) AND (event_type = 3))
 Total runtime: 5.099 ms
```

Query 2:

```
                                                                       QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------------------------
 Sort  (cost=38.97..38.98 rows=1 width=16) (actual time=0.120..0.120 rows=0 loops=1)
   Sort Key: b.created
   Sort Method: quicksort  Memory: 25kB
   ->  Nested Loop  (cost=4.89..38.96 rows=1 width=16) (actual time=0.109..0.109 rows=0 loops=1)
         ->  Bitmap Heap Scan on bookshelves_events e  (cost=4.47..30.52 rows=1 width=13) (actual time=0.079..0.080 rows=1 loops=1)
               Recheck Cond: (username = 'user_42'::text)
               Filter: (event_date ~~ '2024%'::text)
               Rows Removed by Filter: 6
               ->  Bitmap Index Scan on bookshelves_events_user_checkins_idx  (cost=0.00..4.47 rows=7 width=0) (actual time=0.039..0.039 rows=7 loops=1)
                     Index Cond: (username = 'user_42'::text)
         ->  Index Scan using bookshelves_books_pkey on bookshelves_books b  (cost=0.41..8.44 rows=1 width=25) (actual time=0.025..0.025 rows=0 loops=1)
               Index Cond: ((username = 'user_42'::text) AND (work_id = e.work_id))
 Total runtime: 1.072 ms
```

Query 3:
```
                                                                       QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------------------------
 Unique  (cost=26.97..26.98 rows=1 width=78) (actual time=0.065..0.065 rows=1 loops=1)
   ->  Sort  (cost=26.97..26.97 rows=1 width=78) (actual time=0.064..0.064 rows=1 loops=1)
         Sort Key: work_id, updated
         Sort Method: quicksort  Memory: 25kB
         ->  Bitmap Heap Scan on bookshelves_events  (cost=4.49..26.96 rows=1 width=78) (actual time=0.055..0.055 rows=1 loops=1)
               Recheck Cond: ((username = 'user_42'::text) AND (event_type = 3))
               Filter: (event_date ~~ '2024%'::text)
               Rows Removed by Filter: 5
               ->  Bitmap Index Scan on bookshelves_events_user_checkins_idx  (cost=0.00..4.49 rows=6 width=0) (actual time=0.032..0.032 rows=6 loops=1)
                     Index Cond: ((username = 'user_42'::text) AND (event_type = 3))
 Total runtime: 0.108 ms
```


### Technical
<!-- What should be noted about the implementation? -->

### Testing
<!-- Steps for reviewer to reproduce/verify what this PR does/fixes. -->

### Screenshot
<!-- If this PR touches UI, please post evidence (screenshots) of it behaving correctly. -->

### Stakeholders
<!-- @ tag the lead (as labeled on the issue) and other stakeholders -->


<!-- Attribution Disclaimer: By proposing this pull request, I affirm to have made a best-effort and exercised my discretion to make sure relevant sections of this code which substantially leverage code suggestions, code generation, or code snippets from sources (e.g. Stack Overflow, GitHub) have been annotated with basic attribution so reviewers & contributors may have confidence and access to the correct context to evaluate and use this code. -->