@cdrini 2025-03-18
Tested on prod, it shows a solid improvement in EXPLAIN/ANALYZE!


Before:
Total runtime: 91.175 ms
Total runtime: 92.013 ms
Total runtime: 109.075 ms

After:
Total runtime: 0.111 ms
Total runtime: 0.264 ms
Total runtime: 11.682 ms
On prod before:

Query 1:
```
                                                          QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=8141.30..8141.30 rows=1 width=13) (actual time=91.150..91.150 rows=0 loops=1)
   ->  Sort  (cost=8141.30..8141.30 rows=1 width=13) (actual time=91.150..91.150 rows=0 loops=1)
         Sort Key: event_date
         Sort Method: quicksort  Memory: 25kB
         ->  Seq Scan on bookshelves_events  (cost=0.00..8141.28 rows=1 width=13) (actual time=91.142..91.142 rows=0 loops=1)
               Filter: ((work_id = ANY ('{1234,143}'::integer[])) AND (username = 'ScarTissue'::text) AND (event_type = 3))
               Rows Removed by Filter: 221828
 Total runtime: 91.175 ms
```

Query 2:
```
                                                                       QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------------------------
 Sort  (cost=7611.91..7611.92 rows=1 width=16) (actual time=91.941..91.942 rows=8 loops=1)
   Sort Key: b.created
   Sort Method: quicksort  Memory: 25kB
   ->  Nested Loop  (cost=0.56..7611.90 rows=1 width=16) (actual time=44.171..91.895 rows=8 loops=1)
         ->  Seq Scan on bookshelves_events e  (cost=0.00..7601.53 rows=4 width=15) (actual time=42.047..87.095 rows=8 loops=1)
               Filter: ((event_date ~~ '2024%'::text) AND (username = 'ScarTissue'::text))
               Rows Removed by Filter: 221820
         ->  Index Scan using bookshelves_books_pkey on bookshelves_books b  (cost=0.56..2.58 rows=1 width=27) (actual time=0.592..0.593 rows=1 loops=8)
               Index Cond: ((username = 'ScarTissue'::text) AND (work_id = e.work_id))
 Total runtime: 92.013 ms
```

Query 3:

```
                                                          QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------
 Unique  (cost=8141.32..8141.35 rows=2 width=84) (actual time=109.023..109.031 rows=8 loops=1)
   ->  Sort  (cost=8141.32..8141.34 rows=4 width=84) (actual time=109.020..109.021 rows=8 loops=1)
         Sort Key: work_id, updated
         Sort Method: quicksort  Memory: 26kB
         ->  Seq Scan on bookshelves_events  (cost=0.00..8141.28 rows=4 width=84) (actual time=61.809..107.576 rows=8 loops=1)
               Filter: ((event_date ~~ '2024%'::text) AND (username = 'ScarTissue'::text) AND (event_type = 3))
               Rows Removed by Filter: 221820
 Total runtime: 109.075 ms
```

After:

Query 1:

```
                                                                              QUERY PLAN

----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=3.88..3.89 rows=1 width=13) (actual time=0.049..0.049 rows=0 loops=1)
   ->  Sort  (cost=3.88..3.89 rows=1 width=13) (actual time=0.049..0.049 rows=0 loops=1)
         Sort Key: event_date
         Sort Method: quicksort  Memory: 25kB
         ->  Index Scan using bookshelves_events_user_checkins_idx on bookshelves_events  (cost=0.42..3.87 rows=1 width=13) (actual time=0.043..0.043 rows=0 loops=1)
               Index Cond: ((username = 'ScarTissue'::text) AND (work_id = ANY ('{1234,143}'::integer[])) AND (event_type = 3))
 Total runtime: 0.111 ms
(7 rows)
```

Query 2:

```
                                                                                 QUERY PLAN

-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Sort  (cost=16.03..16.04 rows=1 width=16) (actual time=0.230..0.231 rows=8 loops=1)
   Sort Key: b.created
   Sort Method: quicksort  Memory: 25kB
   ->  Nested Loop  (cost=0.98..16.02 rows=1 width=16) (actual time=0.095..0.218 rows=8 loops=1)
         ->  Index Only Scan using bookshelves_events_user_checkins_idx on bookshelves_events e  (cost=0.42..5.65 rows=4 width=15) (actual time=0.033..0.050 rows=8 loops=1)
               Index Cond: ((username = 'ScarTissue'::text) AND (event_date >= '2024'::text) AND (event_date < '2025'::text))
               Filter: (event_date ~~ '2024%'::text)
               Heap Fetches: 8
         ->  Index Scan using bookshelves_books_pkey on bookshelves_books b  (cost=0.56..2.58 rows=1 width=27) (actual time=0.019..0.020 rows=1 loops=8)
               Index Cond: ((username = 'ScarTissue'::text) AND (work_id = e.work_id))
 Total runtime: 0.264 ms
```

Query 3:
```
                                                                              QUERY PLAN

-----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Unique  (cost=5.73..5.75 rows=2 width=84) (actual time=11.627..11.641 rows=8 loops=1)
   ->  Sort  (cost=5.73..5.74 rows=4 width=84) (actual time=11.625..11.626 rows=8 loops=1)
         Sort Key: work_id, updated
         Sort Method: quicksort  Memory: 26kB
         ->  Index Scan using bookshelves_events_user_checkins_idx on bookshelves_events  (cost=0.42..5.69 rows=4 width=84) (actual time=2.606..11.603 rows=8 loops=1)
               Index Cond: ((username = 'ScarTissue'::text) AND (event_type = 3) AND (event_date >= '2024'::text) AND (event_date < '2025'::text))
               Filter: (event_date ~~ '2024%'::text)
 Total runtime: 11.682 ms
```